텀 프로젝트 파일 매니저 시스템 구현 3
- png 파일의 미리보기 위젯 완성
- txt 파일의 내용을 메세지 박스를 통해 미리 볼 수 있다.

+++ term_project.py
@@ -1,8 +1,12 @@
 import os
 import re
+
 from tkinter import *
 from tkinter.ttk import *
 from tkinter.scrolledtext import ScrolledText
+
+import PIL.Image
+from PIL import ImageTk, Image
 import tkinter.messagebox as messagebox
 
 path_list = []
@@ -18,20 +22,53 @@ def set_path():
 
 
 def remove_file():
+    # 리스트 커서를 선택하지 않은 경우
+    if not len(found_result_lstBox.curselection()):
+        messagebox.showinfo("경고", " 리스트를 체크해 주세요")
+        return
     remove_file_path = path_list[found_result_lstBox.curselection()[0]]
     if os.path.exists(remove_file_path):
         os.remove(remove_file_path)
+        path_list.remove(remove_file_path)
+
     found_result_lstBox.delete(found_result_lstBox.curselection()[0], found_result_lstBox.curselection()[0])
     found_result_lstBox.update()
 
 
+def set_preview():
+    if not len(found_result_lstBox.curselection()):
+        messagebox.showinfo("경고", " 리스트를 체크해 주세요")
+        return
+    preview_file_path = path_list[found_result_lstBox.curselection()[0]]
+    if os.path.exists(preview_file_path):
+        if re.search(".png", preview_file_path):
+            print(preview_file_path)
+            prev_img = Image.open(preview_file_path)
+            resized_img = prev_img.resize((160, 160))
+            photo = ImageTk.PhotoImage(resized_img)
+            preview_label.configure(image=photo)
+            preview_label.image = photo
+        elif re.search(".txt", preview_file_path):
+            prev_img = ImageTk.PhotoImage(Image.open(r"C:\Users\OH.S.D\Desktop\preview.png").resize((160, 160)))
+            preview_label.configure(image=prev_img)
+            preview_label.image = prev_img
+            txt_read = open(preview_file_path, 'r')
+            messagebox.showinfo("텍스트 파일 내용", txt_read.read())
+            txt_read.close()
+
+    else:
+        return
+
+
 def search_file(root_path, lst_idx):
+    # 매 검색마다 리스트를 초기화 시켜줌
     if lst_idx == 0:
         path_list.clear()
         found_result_lstBox.delete(0, END)
+
     # 확장자 선택 안했을때 예외처리
     if found_extension.get() == '':
-        messagebox.showerror("에러", "탐색 확장자를 설정해 주시오.")
+        messagebox.showerror("경고", "탐색 확장자를 설정해 주시오.")
         return
 
     if os.path.exists(root_path):
@@ -57,17 +94,15 @@ windows.title("File_Manager")
 # 파일 확장자 설정
 # 파일 확장자 선택을 위한 라디오 버튼
 file_extension_frame = LabelFrame(text="Extension")
-file_extension_frame.pack(side=RIGHT)
+file_extension_frame.pack(side=RIGHT, fill=BOTH)
 
 found_extension = StringVar()
 txt_radioButton = Radiobutton(file_extension_frame, text="txt", value=".txt",
                               variable=found_extension)
-jpg_radioButton = Radiobutton(file_extension_frame, text="jpg", value=".jpg",
-                              variable=found_extension)
+
 png_radioButton = Radiobutton(file_extension_frame, text="png", value=".png",
                               variable=found_extension)
 txt_radioButton.pack()
-jpg_radioButton.pack()
 png_radioButton.pack()
 # -------------------------------------------
 
@@ -80,17 +115,27 @@ input_path.pack(side=LEFT)
 find_button = Button(file_find_frame, text="찾기", command=set_path)
 find_button.pack(side=RIGHT)
 
-# 탐색 결과 리스트
-found_result_lstBox = Listbox(selectmode='single', width=100)
-found_result_lstBox.pack()
+# 탐색 결과 리스트 위젯
+result_frame = Frame()
+
+found_result_lstBox = Listbox(selectmode='single', width=40)
 found_result_lstBox.configure(background="skyblue", foreground="white", font='Aerial')
+found_result_lstBox.pack(side=LEFT, fill=BOTH)
 
-# 파일 삭제 기능
+# 미리보기 위젯
+img = ImageTk.PhotoImage(Image.open(r"C:\Users\OH.S.D\Desktop\preview.png").resize((160, 160)))
+preview_label = Label(result_frame, image=img)
+preview_label.configure(background="WHITE")
+preview_label.pack(side=RIGHT, fill=BOTH)
+result_frame.pack()
+
+# 파일 삭제, 미리보기 버튼 기능
 manipulate_btn_frame = LabelFrame(text="파일 작업")
-delete_btn = Button(text="삭제", command=remove_file)
+delete_btn = Button(manipulate_btn_frame, text="삭제", command=remove_file)
 delete_btn.pack(side=LEFT)
-
-#파일 미리보기 기능
+preview_btn = Button(manipulate_btn_frame, text="미리보기", command=set_preview)
+preview_btn.pack(side=RIGHT)
+manipulate_btn_frame.pack()
 
 
 # 프로그램 종료 키워드
